Patching CAS/Client.php for Pantheon compatability
Patch fixes an error experienced by Pantheon users, which is produced by the return request from Client.php including both port 80 and HTTPS.

The submitted patch on lines 3602-3612 includes a conditional for the detection of a Pantheon environement variable called PANTHEON_ENVIRONMENT which is set globally on WP installations on that service. If that environment variable is detected it bypasses the offending code and sends the return request without the port number.